Give the registry-items tree its own module
`RegistryItems` lived inside core's `Registry` module while the
queries over it (`collectFunctionNames`, `collectMiddlewareKeys`) hid
as private helpers in `GroupImpl`. Bring the tree type and its queries
together in a server `RegistryItems` module — the queries narrow leaves
with `FunctionRegistryItem`/`MiddlewareRegistryItem` predicates, so the
module must live where those do — and name the queries after what they
return: `functionNames` and `middlewareKeys`.

The `Registry` reference follows along into `@confect/server` (its own
module), since the tree type was its only reason to sit in core and
nothing in core uses it; the CLI already depends on the server package
and now imports the same tag from there. Its key becomes
`@confect/server/Registry` to match its new home.
