+++ packages/core/src/index.ts
@@ -10,7 +10,6 @@ export * as PaginationOptions from "./PaginationOptions";
 export * as PaginationResult from "./PaginationResult";
 export * as Ref from "./Ref";
 export * as Refs from "./Refs";
-export * as Registry from "./Registry";
 export * as RuntimeAndFunctionType from "./RuntimeAndFunctionType";
 export * as Spec from "./Spec";
 export * as SystemFields from "./SystemFields";
