+++ packages/server/src/MiddlewareImpl.ts
@@ -1,7 +1,7 @@
 import type * as FunctionSpec from "@confect/core/FunctionSpec";
 import type * as GroupSpec from "@confect/core/GroupSpec";
 import * as MiddlewareSpec from "@confect/core/MiddlewareSpec";
-import * as Registry from "@confect/core/Registry";
+import * as Registry from "./Registry";
 import type { FunctionType } from "convex/server";
 import * as Context from "effect/Context";
 import * as Effect from "effect/Effect";
